fix(babel-plugin-docz): add libundler to build using lerna
