--- packages/babel-plugin-docz/.gitignore
@@ -1,2 +1,0 @@
-.vscode
-/node_modules
