--- packages/babel-plugin-docz/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "printWidth": 80,
-  "singleQuote": true
-}
