+++ packages/babel-plugin-docz/package.json
@@ -2,13 +2,25 @@
   "name": "babel-plugin-docz",
   "version": "0.12.2",
   "description": "Add superpowers to docz",
-  "main": "lib/index.js",
-  "repository": "https://github.com/mpivaa/babel-plugin-docz",
-  "author": "Marcelo Piva <m.pivaa@gmail.com>",
   "license": "MIT",
+  "author": "Marcelo Piva <m.pivaa@gmail.com>",
+  "main": "dist/index.js",
+  "umd:main": "dist/index.umd.js",
+  "module": "dist/index.m.js",
+  "typings": "dist/index.d.ts",
+  "source": "src/index.ts",
+  "files": [
+    "dist/",
+    "package.json",
+    "README.md"
+  ],
   "scripts": {
-    "build": "babel src/index.js -o lib/index.js",
-    "dev": "babel src/index.js -o lib/index.js --watch"
+    "dev": "libundler watch --ts -e all",
+    "build": "libundler build --ts -e all --c",
+    "fix": "run-s fix:*",
+    "fix:prettier": "prettier \"src/**/*.{ts,tsx}\" --write",
+    "fix:tslint": "tslint --fix --project .",
+    "tslint": "tslint --project ."
   },
   "dependencies": {
     "@babel/cli": "^7.1.2",
