+++ packages/babel-plugin-docz/src/types.d.ts
@@ -1,0 +1,1 @@
+declare module 'babel-template'
